+++ CHANGELOG.md
@@ -23,11 +23,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [rust-0.2.0] - Unreleased
 
+### Removed
+
+- The `NoOp` update handler is gone in favor of `()`
+
 ### Added
 
 - `DataSource::Polars` to support direct loading of Polars' DataFrame.
 - `Engine::del_column` to delete columns
 - `Codebook::from_df` to create a codebook from a polars DataFrame
+- `()` now implements `UpdateHandler` (replaces `NoOp`)
 
 ### Fixed